requirements.txt
python-telegram-bot==20.8
google-generativeai==0.7.2
python-dotenv==1.0.0
